refactor(stage): move reload-decision helpers to ndi_reload_guard (keep ndi_watchdog under 1000-line cap)
The #422 test-only bypass pushed ndi_watchdog.rs over the 1000-prod-line hard cap
(1033). Move the two pure reload-decision helpers — should_escalate_reload and
reload_skip_healthz_from_url — to ndi_reload_guard.rs (their natural home: the
reload-decision module). ndi_watchdog drops to 990 prod lines. No behavior change;
callers use super::ndi_reload_guard::. #418 still tracks the remaining over-target
size for a fuller split post-event.
